src/agent-app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai==1.50.0
azure-search-documents==11.6.0b4
azure-identity==1.18.0
azure-monitor-opentelemetry==1.6.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
httpx==0.27.0
python-dotenv==1.0.1
setuptools>=70.0.0,<71